+++ ExplorerXP/Viewers/RecBinViewer.h
@@ -1,4 +1,4 @@
-/* Copyright 2002-2020 Nikolay Avrionov. All Rights Reserved.
+/* Copyright 2002-2021 Nikolay Avrionov. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -73,7 +73,7 @@ public:
 
 	const TCHAR*  GetName (int iRow) ;
 	//int GetColumnCount () { return m_Headers.size ();};
-	int GetCount () { return m_List.size ();}
+	int GetCount () { return static_cast<int>(m_List.size ());}
 	bool GetObjectSize (int iRow, ULONGLONG &size, ULONGLONG &sizeOnDisk);
 
 	bool CanPaste () { return false;};
